realtime-voice-server/requirements.txt
# Realtime Voice Server Dependencies
websockets>=14.0
python-dotenv>=1.0.0
openai>=1.3.7
pandas>=1.5.0
asyncio
uuid

# Enhanced RAG Service Dependencies
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langchain-core>=0.3.34
langchain-community>=0.0.2
pinecone-client>=2.2.0

# Additional dependencies for enhanced functionality
requests>=2.31.0
numpy>=1.25.0 